Fix: Refactor JSLL and chisel
