+++ rules/chisel/.eslintignore
@@ -1,0 +1,3 @@
+coverage
+dist
+node_modules
