+++ rules/chisel/.npmrc
