+++ rules/chisel/.sonarwhalrc
@@ -1,11 +1,12 @@
 {
     "connector": {
-        "name": "chrome",
+        "name": "jsdom",
         "options": {
             "waitFor": 1000
         }
     },
     "formatters": "stylish",
+    "parsers": [],
     "rulesTimeout": 120000,
     "rules": {
         "chisel": "warning"
