+++ rules/chisel/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 1.0.0 (February 8, 2018)
+
+✨
