+++ rules/chisel/package.json
@@ -12,14 +12,15 @@
     "timeout": "1m"
   },
   "dependencies": {
-    "@mskeros/chisel": "^1.130.1"
+    "@mskeros/chisel": "^1.132.3",
+    "@sonarwhal/formatter-stylish": "^1.0.1"
   },
   "description": "sonarwhal rule to assess the accessibility of a web site using Microsoft's Chisel.",
   "devDependencies": {
     "@types/debug": "0.0.30",
     "@types/node": "8.0.14",
     "async-retry": "^1.1.4",
-    "ava": "^0.24.0",
+    "ava": "^0.25.0",
     "babel-cli": "^6.26.0",
     "babel-plugin-istanbul": "^4.1.5",
     "babel-preset-es2017": "^6.22.0",
@@ -31,9 +32,10 @@
     "express": "^4.16.2",
     "markdownlint-cli": "^0.6.0",
     "npm-run-all": "^4.1.2",
+    "nyc": "^11.4.1",
     "on-headers": "^1.0.1",
     "rimraf": "^2.6.2",
-    "sonarwhal": "^0.25.0",
+    "sonarwhal": "^0.28.2",
     "typescript": "^2.6.2",
     "typescript-eslint-parser": "^12.0.0"
   },
@@ -55,20 +57,23 @@
   "main": "dist/src/index.js",
   "name": "sonarwhal-rule-chisel",
   "peerDependencies": {
-    "sonarwhal": "^0.25.0"
+    "sonarwhal": "^0.28.2"
   },
   "repository": "Microsoft/sonarwhal-microsoft-rules",
   "scripts": {
     "build": "npm run clean && npm-run-all build:*",
+    "build-release": "npm run clean && npm run build:assets && tsc --inlineSourceMap false --removeComments true",
     "build:assets": "cpx \"./{src,tests}/**/{!(*.ts),.!(ts)}\" dist",
-    "build:ts": "tsc --outDir \"./dist\"",
+    "build:ts": "tsc",
     "clean": "rimraf dist",
-    "test": "npm run lint && npm run build && ava",
     "lint": "npm-run-all lint:*",
-    "lint:js": "eslint --ext md --ext ts --ignore-pattern dist .",
-    "lint:md": "markdownlint README.md",
+    "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path .eslintignore --report-unused-disable-directives",
+    "lint:md": "markdownlint --ignore CHANGELOG.md *.md",
+    "test": "npm run lint && npm run build && nyc ava",
+    "watch": "npm run build && npm-run-all --parallel -c watch:*",
+    "watch:assets": "npm run build:assets -- -w --no-initial",
+    "watch:test": "ava --watch",
     "watch:ts": "npm run build:ts -- --watch",
-    "init": "npm install && npm run build",
     "sonarwhal": "node node_modules/sonarwhal/dist/src/bin/sonarwhal.js"
   },
   "version": "0.1.0"
