+++ rules/jsll/package.json
@@ -12,31 +12,32 @@
     "timeout": "1m"
   },
   "dependencies": {
+    "@sonarwhal/formatter-stylish": "^1.0.1",
+    "@sonarwhal/parser-javascript": "^1.0.1",
     "locale-code": "^1.1.1"
   },
   "description": "sonarwhal rule to validate Microsoft's JSLL configuration.",
   "devDependencies": {
     "@types/debug": "0.0.30",
-    "@types/node": "8.0.14",
-    "async-retry": "^1.1.4",
-    "ava": "^0.24.0",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-istanbul": "^4.1.5",
-    "babel-preset-es2017": "^6.22.0",
-    "babel-register": "^6.26.0",
+    "@types/node": "^9.4.6",
+    "async-retry": "^1.2.1",
+    "ava": "^0.25.0",
     "cpx": "^1.5.0",
-    "eslint": "^4.15.0",
-    "eslint-plugin-markdown": "^1.0.0-beta.6",
-    "eslint-plugin-typescript": "^0.8.1",
+    "eslint": "^4.18.2",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-markdown": "^1.0.0-beta.7",
+    "eslint-plugin-typescript": "^0.9.0",
     "express": "^4.16.2",
     "lodash.clonedeep": "^4.5.0",
-    "markdownlint-cli": "^0.6.0",
+    "markdownlint-cli": "^0.7.0",
+    "npm-link-check": "^2.0.0",
     "npm-run-all": "^4.1.2",
+    "nyc": "^11.4.1",
     "on-headers": "^1.0.1",
     "rimraf": "^2.6.2",
-    "sonarwhal": "^0.25.0",
+    "sonarwhal": "^0.28.2",
     "typescript": "^2.6.2",
-    "typescript-eslint-parser": "^12.0.0"
+    "typescript-eslint-parser": "^14.0.0"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -47,29 +48,31 @@
   ],
   "homepage": "https://sonarwhal.com",
   "keywords": [
-    "jsll",
-    "rule",
     "sonarwhal",
-    "sonarwhal-rule"
+    "sonarwhal-rule",
+    "jsll",
+    "jsll-rule"
   ],
   "license": "Apache-2.0",
   "main": "dist/src/index.js",
   "name": "sonarwhal-rule-jsll",
   "peerDependencies": {
-    "sonarwhal": "^0.25.0"
+    "sonarwhal": "^0.28.2"
   },
   "scripts": {
-    "ava": "ava",
     "build": "npm run clean && npm-run-all build:*",
+    "build-release": "npm run clean && npm run build:assets && tsc --inlineSourceMap false --removeComments true",
     "build:assets": "cpx \"./{src,tests}/**/{!(*.ts),.!(ts)}\" dist",
-    "build:ts": "tsc --outDir \"./dist\"",
+    "build:ts": "tsc",
     "clean": "rimraf dist",
-    "test": "npm run lint && npm run build && ava",
     "lint": "npm-run-all lint:*",
-    "lint:js": "eslint --ext md --ext ts --ignore-pattern dist .",
-    "lint:md": "markdownlint README.md",
+    "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path .eslintignore --report-unused-disable-directives",
+    "lint:md": "markdownlint --ignore CHANGELOG.md *.md",
+    "test": "npm run lint && npm run build && nyc ava",
+    "watch": "npm run build && npm-run-all --parallel -c watch:*",
+    "watch:assets": "npm run build:assets -- -w --no-initial",
+    "watch:test": "ava --watch",
     "watch:ts": "npm run build:ts -- --watch",
-    "init": "npm install && npm run build",
     "sonarwhal": "node node_modules/sonarwhal/dist/src/bin/sonarwhal.js"
   },
   "repository": "Microsoft/sonarwhal-microsoft-rules",
